requirements.txt
fastapi
uvicorn
opencv-python
numpy
Pillow
torch
facenet-pytorch
yagmail
ultralytics
httpx
pydantic
